modulo4/consolidacion/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
python-dotenv==1.0.0
openpyxl==3.1.2
pandas==2.0.3
Pillow==10.0.0
reportlab==4.0.4
xhtml2pdf==0.2.13